remove pip install method
No new packages can be added to the internal store using the pip method.
Previously installed packages continue to be usable for now.

Parsing of requirements.txt files becomes obsolete with this.
The cases where the `Requirement` type was used during installation
were replaced with `Distribution`.

The following function was moved:
Requirement::from_filename()
-> Distribution::from_package_name()

along with its unit test.
