+++ src/internal_store.rs
@@ -1,7 +1,6 @@
 use eyre::{ensure, eyre, WrapErr};
 use fs_err::File;
 
-use crate::python::wheel::is_path_of_executable;
 use crate::python::{
     print_error_missing_file_in_record, records, Distribution, DistributionHash, EntryPoint,
     FileHash, PythonVersion,
@@ -12,9 +11,8 @@ use crate::venv::{
 };
 use crate::{
     delete_virtpy_backing, is_not_found, package_info_from_dist_info_dirname,
-    python::requirements::Requirement,
     python::wheel::{RecordEntry, WheelRecord},
-    remove_leading_parent_dirs, EResult, Options, Path, PathBuf, ProjectDirs, INVALID_UTF8_PATH,
+    EResult, Options, Path, PathBuf, ProjectDirs, INVALID_UTF8_PATH,
 };
 use std::path::Path as StdPath;
 use std::{
@@ -364,81 +362,6 @@ fn unused_package_files(proj_dirs: &ProjectDirs) -> impl Iterator<Item = PathBuf
         .map(move |(file, _)| package_files.join(file))
 }
 
-fn register_distribution_files(
-    proj_dirs: &ProjectDirs,
-    install_folder: &Path,
-    distribution: &Distribution,
-    stored_distributions: &mut HashMap<DistributionHash, StoredDistribution>,
-    options: crate::Options,
-) -> EResult<()> {
-    let dist_info_foldername = distribution.dist_info_name();
-    let src_dist_info = install_folder.join(&dist_info_foldername);
-
-    let stored_distrib = StoredDistribution {
-        distribution: distribution.clone(),
-        installed_via: StoredDistributionType::FromPip,
-    };
-    let dst_dist_info = proj_dirs
-        .dist_infos()
-        .join(&stored_distrib.distribution.as_csv());
-
-    // let use_move = can_move_files(&proj_dirs.package_files(), install_folder).unwrap_or(false);
-    let use_move = true;
-
-    if dst_dist_info.exists() {
-        // add it here, because it may have been installed by a different
-        // python version. In that case, the current python version's list
-        // may be missing this distribution.
-        stored_distributions.insert(distribution.sha.clone(), stored_distrib);
-        return Ok(());
-    }
-    if options.verbose >= 1 {
-        println!(
-            "Adding {} {} to central store.",
-            distribution.name, distribution.version
-        );
-    }
-
-    for file in records(&src_dist_info.join("RECORD"))
-        .wrap_err("couldn't find dist-info/RECORD")?
-        .map(Result::unwrap)
-    {
-        // Sanity check. We're not caching compiled code so pip is told not to compile python code.
-        // If this folder exists, something went wrong.
-        debug_assert!(file.path.iter().all(|part| part != "__pycache__"));
-
-        let path = remove_leading_parent_dirs(&file.path).unwrap_or_else(std::convert::identity);
-        if is_path_of_executable(path) {
-            // executables generated by pip depend either on the global python
-            // or the venv they were generated for. They can't be symlinked into a virtpy
-            // but have to be generated on demand.
-            continue;
-        }
-
-        debug_assert_ne!(file.hash, FileHash("".to_owned()));
-
-        let src = install_folder.join(path);
-        let dest = proj_dirs.package_file(&file.hash);
-        if options.verbose >= 2 {
-            println!("    copying {} to {}", src, dest);
-        }
-
-        let res = move_file(&src, &dest, use_move);
-        match &res {
-            Err(err) if is_not_found(err) => {
-                print_error_missing_file_in_record(distribution, file.path.as_ref())
-            }
-            _ => {
-                res.unwrap();
-            }
-        };
-    }
-
-    copy_directory(&src_dist_info, &dst_dist_info, use_move);
-    stored_distributions.insert(distribution.sha.clone(), stored_distrib);
-    Ok(())
-}
-
 fn register_distribution_files_of_wheel(
     proj_dirs: &ProjectDirs,
     install_folder: &Path,
@@ -773,59 +696,6 @@ impl std::ops::DerefMut for FileLockGuard {
     }
 }
 
-pub(crate) fn register_new_distributions(
-    options: Options,
-    new_distribs: Vec<Distribution>,
-    n_distribs_requested: usize,
-    proj_dirs: &ProjectDirs,
-    pip_log: String,
-    python_version: PythonVersion,
-    tmp_dir: tempdir::TempDir,
-) -> EResult<()> {
-    if options.verbose >= 1 && new_distribs.len() != n_distribs_requested {
-        // either an error or a sign that the filters in new_dependencies()
-        // need to be improved
-        println!(
-            "Only found {} of {} distributions",
-            new_distribs.len(),
-            n_distribs_requested
-        );
-
-        let _ = fs_err::write(proj_dirs.data().join("pip.log"), pip_log);
-    }
-    if options.verbose >= 2 {
-        for distrib in new_distribs.iter() {
-            println!(
-                "    New distribution: {}=={}, {}",
-                distrib.name, distrib.version, distrib.sha.0
-            );
-        }
-    }
-    let mut all_stored_distributions = StoredDistributions::load(proj_dirs)?;
-    let stored_distributions = all_stored_distributions
-        .0
-        .entry(python_version.as_string_without_patch())
-        .or_default();
-    for distrib in new_distribs {
-        register_distribution_files(
-            proj_dirs,
-            tmp_dir.path().try_into().expect(INVALID_UTF8_PATH),
-            &distrib,
-            stored_distributions,
-            options,
-        )
-        .wrap_err_with(|| {
-            eyre!(
-                "failed to add distribution files for {} {}",
-                distrib.name,
-                distrib.version
-            )
-        })?;
-    }
-    all_stored_distributions.save()?;
-    Ok(())
-}
-
 // Usable only for our own installation from wheel files
 pub(crate) fn register_new_distribution(
     options: Options,
@@ -863,35 +733,6 @@ pub(crate) fn register_new_distribution(
     Ok(())
 }
 
-pub(crate) fn new_dependencies(
-    requirements: &[Requirement],
-    proj_dirs: &ProjectDirs,
-    python_version: PythonVersion,
-) -> EResult<Vec<Requirement>> {
-    let stored_distributions = StoredDistributions::load(proj_dirs)?;
-    let existing_deps = match stored_distributions
-        .0
-        .get(&python_version.as_string_without_patch())
-    {
-        Some(deps) => deps,
-        None => return Ok(requirements.to_owned()),
-    };
-
-    Ok(requirements
-        .iter()
-        .filter(|req| {
-            (req.marker
-                .as_ref()
-                .map_or(true, |cond| cond.matches_system()))
-                && !req
-                    .available_hashes
-                    .iter()
-                    .any(|hash| existing_deps.contains_key(hash))
-        })
-        .cloned()
-        .collect::<Vec<_>>())
-}
-
 pub(crate) fn wheel_is_already_registered(
     wheel_hash: DistributionHash,
     proj_dirs: &ProjectDirs,
@@ -902,24 +743,6 @@ pub(crate) fn wheel_is_already_registered(
         .0
         .get(&python_version.as_string_without_patch())
         .map_or(false, |deps| deps.contains_key(&wheel_hash)))
-}
-
-fn copy_directory(from: impl AsRef<StdPath>, to: impl AsRef<StdPath>, use_move: bool) {
-    _copy_directory(from.as_ref(), to.as_ref(), use_move)
-}
-
-fn _copy_directory(from: &StdPath, to: &StdPath, use_move: bool) {
-    for dir_entry in walkdir::WalkDir::new(from) {
-        let dir_entry = dir_entry.unwrap();
-        let path = dir_entry.path();
-        let subpath = path.strip_prefix(from).unwrap();
-        let target_path = to.join(&subpath);
-        if dir_entry.file_type().is_dir() {
-            fs_err::create_dir(target_path).unwrap();
-        } else {
-            move_file(path, &target_path, use_move).unwrap();
-        }
-    }
 }
 
 // fn dist_info_dirname(name: &str, version: &str, hash: &DistributionHash) -> String {
